1. add code_util 2. better encapsulation
